Replace draw queue with new generic draw function
In this third iteration the game manager will have a draw function that
takes the shader to draw, a vertex array and textures, and tha target
framebuffer.
A getVertArray function has been added to the Mesh object so that it can
use the new generic draw function.

The plan is to have a configurable render pipline defined and for the
game manager to run that pipeline each frame.