Impliment lua binding for parameter type
The parameter datatype now has a function for converting a string value
pair to a parameter, and a given parameter to a value in the lua
context. The shader lua type uses these functions to edit persistant
parameters.

A vector or matrix is returned as a table, with vectors as an array and
matricies as a nested table in row major order.