Added Lua binding for text rendering
The font user datatype has a draw metamethod that takes the shader,
position, scale, and color like the Font class method as well as the
target framebuffer.

It's best to use some sort of projection matrix for the shader as the
position of the text vertecies are not normilized to screen coordinates.